requirements.txt
# The libraries used by your code should be listed here
# See https://nhsd-git.digital.nhs.uk/data-services/analytics-service/iuod/rap-community-of-practice/-/blob/master/python/project-structure-and-packaging.md

-e .

# Libraries used for running the NCMP pipeline
# Package versions frozen as of 30JUN2022
# Python version = 3.7.*

# Data manipulation
numpy==1.21.5
pandas==1.3.5

# Excel output
xlwings==0.24.9
openpyxl==3.0.09

# SQL
sqlalchemy==1.4.32
pyodbc==4.0.32

# Testing
pytest==6.2.5
pytest-html==3.1.1


# Dependencies of the above packages
black==22.1.0
cffi==1.15.0
debugpy==1.5.1
executing==0.8.2
importlib-metadata==4.10.1
importlib-resources==5.4.0
iniconfig==1.1.1
ipykernel==6.9.0
ipython==7.32.0
ipython-genutils==0.2.0
jedi==0.14.1
jsonschema==4.4.0
jupyter-client==7.1.2
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
kiwisolver==1.3.2
MarkupSafe==2.0.1
mistune==0.8.4
mypy-extensions==0.4.3
nbclient==0.5.10
nbconvert==6.4.1
nbformat==5.1.3
nest-asyncio==1.5.4
packaging==21.3
pandocfilters==1.5.0
pathspec==0.9.0
patsy==0.5.2
pickleshare==0.7.5
Pillow==9.0.1
pip==21.2.4
py==1.11.0
pyparsing==3.0.7
pyrsistent==0.18.1
pytest-metadata==1.11.0
python-dateutil==2.8.2
pywin32==303
pywinpty==2.0.2
pyzmq==22.3.0
Send2Trash==1.8.0
setuptools==59.8.0
simplegeneric==0.8.1
testpath==0.5.0
tornado==6.1
traitlets==5.1.1
typed-ast==1.5.2
typing_extensions==4.0.1
tzdata==2021.5
tzlocal==4.1
webencodings==0.5.1
zipp==3.7.0
